requirements.txt
numpy==2.1.3
keras==3.10.0
tensorflow==2.19.0
pillow==10.4.0
streamlit==1.46.1